twitter_transcript/requirements.txt
yt-dlp
openai-whisper
torch
setuptools-rust
